requirements.txt
torch==2.3.0
torchvision==0.18.0
torchaudio==2.3.0
sentence-transformers==2.6.0
numpy==1.26.4
pandas>=2.1,<3.0
scikit-learn>=1.3,<2.0
gradio==4.44.0
chromadb==0.4.24
faiss-cpu==1.7.4
tqdm>=4.66
pdfplumber>=0.11,<1.0
plotly>=5.18,<6.0
